Example: GTK Button gets stuck when pressed mouse moves away
If a GTK button is pressed and the mouse moves to another layer
while being held pressed the button is not properly released.
This does not occur if gtk-layer-shell isn't used, as can be easily
tested by removing the specific code from the example.

The easiest way to experience this is by setting "mouse_warping container"
on sway and have the button open a new window. This is the example used.
However, it should be pointed out that the "forkExec" function is not
interfering here. The bug can also be reproduced by just quickly moving
the mouse while pressed.